+++ assets/js/phoenix_live_view/hooks.js
@@ -1,4 +1,6 @@
 import {
+  PHX_ACTIVE_ENTRY_REFS,
+  PHX_LIVE_FILE_UPDATED,
   PHX_PREFLIGHTED_REFS,
   PHX_UPLOAD_REF
 } from "./constants"
@@ -7,6 +9,8 @@ import LiveUploader from "./live_uploader"
 
 let Hooks = {
   LiveFileUpload: {
+    activeRefs(){ return this.el.getAttribute(PHX_ACTIVE_ENTRY_REFS) },
+
     preflightedRefs(){ return this.el.getAttribute(PHX_PREFLIGHTED_REFS) },
 
     mounted(){ this.preflightedWas = this.preflightedRefs() },
@@ -19,6 +23,9 @@ let Hooks = {
           this.__view.cancelSubmit(this.el.form)
         }
       }
+
+      if(this.activeRefs() === ""){ this.el.value = null }
+      this.el.dispatchEvent(new CustomEvent(PHX_LIVE_FILE_UPDATED))
     }
   },
 
@@ -26,7 +33,13 @@ let Hooks = {
     mounted(){
       this.ref = this.el.getAttribute("data-phx-entry-ref")
       this.inputEl = document.getElementById(this.el.getAttribute(PHX_UPLOAD_REF))
-      LiveUploader.getEntryDataURL(this.inputEl, this.ref, url => this.el.src = url)
+      LiveUploader.getEntryDataURL(this.inputEl, this.ref, url => {
+        this.url = url
+        this.el.src = url
+      })
+    },
+    destroyed(){
+      URL.revokeObjectURL(this.url)
     }
   }
 }