+++ assets/js/phoenix_live_view/live_uploader.js
@@ -25,9 +25,7 @@ export default class LiveUploader {
 
   static getEntryDataURL(inputEl, ref, callback){
     let file = this.activeFiles(inputEl).find(file => this.genFileRef(file) === ref)
-    let reader = new FileReader()
-    reader.onload = (e) => callback(e.target.result)
-    reader.readAsDataURL(file)
+    callback(URL.createObjectURL(file))
   }
 
   static hasUploadsInProgress(formEl){
@@ -41,14 +39,14 @@ export default class LiveUploader {
   }
 
   static serializeUploads(inputEl){
-    let files = this.activeFiles(inputEl, "serialize")
+    let files = this.activeFiles(inputEl)
     let fileData = {}
     files.forEach(file => {
       let entry = {path: inputEl.name}
       let uploadRef = inputEl.getAttribute(PHX_UPLOAD_REF)
       fileData[uploadRef] = fileData[uploadRef] || []
       entry.ref = this.genFileRef(file)
-      entry.name = file.name
+      entry.name = file.name || entry.ref
       entry.type = file.type
       entry.size = file.size
       fileData[uploadRef].push(entry)