+++ lib/phoenix_live_view/helpers.ex
@@ -771,16 +771,17 @@ defmodule Phoenix.LiveView.Helpers do
       <% end %>
   """
   def live_img_preview(%Phoenix.LiveView.UploadEntry{ref: ref} = entry, opts \\ []) do
-    opts =
-      Keyword.merge(opts,
-        id: "phx-preview-#{ref}",
-        data_phx_upload_ref: entry.upload_ref,
-        data_phx_entry_ref: ref,
-        data_phx_hook: "Phoenix.LiveImgPreview",
-        data_phx_update: "ignore"
-      )
+    attrs = Keyword.merge(opts, [
+      id: "phx-preview-#{ref}",
+      data_phx_upload_ref: entry.upload_ref,
+      data_phx_entry_ref: ref,
+      data_phx_hook: "Phoenix.LiveImgPreview",
+      data_phx_update: "ignore"
+    ])
 
-    Phoenix.HTML.Tag.content_tag(:img, "", opts)
+    assigns = LiveView.assign(%{__changed__: nil}, attrs: attrs)
+
+    ~H"<img {@attrs}/>"
   end
 
   @doc """
@@ -881,7 +882,7 @@ defmodule Phoenix.LiveView.Helpers do
 
   The `:for` assign is the form's source data and the optional `:action`
   assign can be provided for the form's action. Additionally accepts
-  the same options as `Phoenix.HTML.form_for/4` as optional assigns:
+  the same options as `Phoenix.HTML.Form.form_for/4` as optional assigns:
 
     * `:as` - the server side parameter in which all params for this
       form will be collected (i.e. `as: :user_params` would mean all fields