+++ test/support/layout_view.ex
@@ -12,49 +12,49 @@ defmodule Phoenix.LiveViewTest.LayoutView do
   end
 
   def render("live.html", assigns) do
-    ~L"""
+    ~H"""
     LIVELAYOUTSTART-<%= @val %>-<%= @inner_content %>-LIVELAYOUTEND
     """
   end
 
   def render("live-override.html", assigns) do
-    ~L"""
+    ~H"""
     LIVEOVERRIDESTART-<%= @val %>-<%= @inner_content %>-LIVEOVERRIDEEND
     """
   end
 
   def render("widget.html", assigns) do
-    ~L"""
+    ~H"""
     WIDGET:<%= live_render(@conn, Phoenix.LiveViewTest.ClockLive) %>
     """
   end
 
   def render("with-function-component.html", assigns) do
-    ~L"""
+    ~H"""
     RENDER:<%= component(&Phoenix.LiveViewTest.FunctionComponent.render/1, value: "from component") %>
     """
   end
 
   def render("layout-with-function-component.html", assigns) do
-    ~L"""
+    ~H"""
     LAYOUT:<%= component(&Phoenix.LiveViewTest.FunctionComponent.render/1, value: "from layout") %>
     <%= @inner_content %>
     """
   end
 
   def render("hello.html", assigns) do
-    ~L"""
+    ~H"""
     Hello
     """
   end
 
   def render("styled.html", assigns) do
-    ~L"""
+    ~H"""
     <html>
       <head>
         <title>Styled</title>
         <link rel="stylesheet" href="/css/custom.css"/>
-        <link rel="stylesheet" href="<%= Routes.static_path(@conn, "/css/app.css") %>"/>
+        <link rel="stylesheet" href={Routes.static_path(@conn, "/css/app.css")}/>
         <link rel="stylesheet" href="//example.com/a.css"/>
         <link rel="stylesheet" href="https://example.com/b.css"/>
         <style>body { background-color: #eee; }</style>